# marshal: don't escape quotes unnecessarily